+++ lib/src/widgets/link_preview_card.dart
@@ -61,7 +61,7 @@ class LinkPreviewCard extends StatelessWidget {
         onTap: () => onTap != null
             ? onTap?.call(data)
             : handleNavigation
-                ? _launchUrl(data.originalUrl)
+                ? _launchUrl(data.originalUrl.toString())
                 : null,
         child: Column(
           crossAxisAlignment: CrossAxisAlignment.start,